TODO: exceptions, apicontrollers, services

+++ backend/src/main/java/RoomReservationSystem/service/impl/BuildingServiceImpl.java
@@ -1,6 +1,8 @@
 package RoomReservationSystem.service.impl;
 
 import RoomReservationSystem.dto.BuildingDTO;
+import RoomReservationSystem.exception.BuildingAlredyExistsException;
+import RoomReservationSystem.exception.BuildingNotExistsException;
 import RoomReservationSystem.model.Building;
 import RoomReservationSystem.repository.BuildingRepository;
 import RoomReservationSystem.service.BuildingService;
@@ -21,53 +23,62 @@ import org.springframework.stereotype.Service;
 public class BuildingServiceImpl implements BuildingService {
     @Autowired
     private BuildingRepository buildingRepository;
-    
-    /**
-     * Az épület törlésére szolgáló függvény
-     * @param   building    A törlendő épület
-     */
-    @Override
-    public void delete(Building building) {
-        buildingRepository.delete(building);
-    }
 
     /**
      * Az épület név alapján történő törlésére szolgáló függvény
      * @param   name    A törlendő épület neve
+     * @throws RoomReservationSystem.exception.BuildingNotExistsException
      */
     @Override
-    public void deleteByName(String name) {
-        buildingRepository.deleteByName(name);
+    public void deleteByName(String name) throws BuildingNotExistsException{
+        if (buildingRepository.findByName(name) == null)
+            throw new BuildingNotExistsException("Nincs ilyen névvel rendelkező épület!");
+        else
+            buildingRepository.deleteByName(name);
     }
 
     /**
      * Az épület mentésére szolgáló függvény
      * @param   building    Az épület, amit menteni szeretnénk
      * @return              A mentett épület
+     * @throws RoomReservationSystem.exception.BuildingAlredyExistsException
      */
     @Override
-    public Building save(Building building) {
-        return buildingRepository.save(building);
+    public Building save(Building building) throws BuildingAlredyExistsException{
+        if (buildingRepository.findByName(building.getName()) != null)
+            throw new BuildingAlredyExistsException("Már létezik névvel rendelkező épület!");
+        else
+            return buildingRepository.save(building);
     }
 
     /**
      * A név alapján történő keresést megvalósító függvény
      * @param   name    Az épület neve  
      * @return          Az épület ha létezik, null egyébként
+     * @throws RoomReservationSystem.exception.BuildingNotExistsException
      */
     @Override
-    public Building findByName(String name) {
-        return buildingRepository.findByName(name);
+    public Building findByName(String name) throws BuildingNotExistsException {
+        Building building = buildingRepository.findByName(name);
+        if (building == null)
+            throw new BuildingNotExistsException("Nem létezik ilyen névvel rendelkező épület!");
+        else
+            return building;
     }
     
     /**
      * Az id alapján történő keresést megvalósító függvény
      * @param   id  Az épület id-ja
      * @return      Az épület ha létezik, null egyébként
+     * @throws RoomReservationSystem.exception.BuildingNotExistsException
      */
     @Override
-    public Building findById(int id) {
-        return buildingRepository.findById(id);
+    public Building findById(int id) throws BuildingNotExistsException{
+        Building building = buildingRepository.findById(id);
+        if (building == null)
+            throw new BuildingNotExistsException("Nem létezik ilyen azonosítóval rendelkező épület!");
+        else
+            return building;
     }
 
     /**
